Added tests for progress reporting #1

+++ Source/src/RtwoDtwo.IO.Compression/RtwoDtwo/IO/Compression/StreamExtensions.cs
@@ -11,7 +11,7 @@ namespace RtwoDtwo.IO.Compression
 		{
 			if (stream.CanSeek)
 			{
-				return stream.Position / stream.Length;
+				return stream.Position / (double)stream.Length;
 			}
 
 			return null;
